+++ src/lib.rs
@@ -512,6 +512,7 @@ pub mod metadata;
 pub mod metadata_fdis;
 pub mod modular;
 pub mod modular_fdis;
+pub mod multi_pass_decode;
 pub mod non_zeros_grid;
 pub mod pass_group;
 pub mod pass_group_hf;